Move dip_dip_hab to prob_rule.py
